select responses according to task_uuid and req_seq_num
Former-commit-id: 56120794e9494fe5ebaf239d0afbc0ac9ee25b5b

+++ management/server/broker/manager-command/src/main/java/org/safehaus/kiskis/mgmt/server/command/CommandManager.java
@@ -147,14 +147,14 @@ public class CommandManager implements CommandManagerInterface, BrokerListener {
 
     public List<Response> getResponses() {
         try {
-            return persistenceCommand.getResponses("taskuuid", 1);
+            return persistenceCommand.getResponses("taskuuid", 1l);
         } catch (Exception ex) {
             LOG.log(Level.SEVERE, "Error in getResponses", ex);
         }
         return null;
     }
 
-    public Response getResponse(String taskuuid, Integer requestSequenceNumber) {
+    public Response getResponse(String taskuuid, Long requestSequenceNumber) {
         Response response = null;
         try {
             List<Response> list = persistenceCommand.getResponses(taskuuid, requestSequenceNumber);
